SYNC-1569 Regenerate

+++ modules/apps/sync/sync-api/src/main/java/com/liferay/sync/service/SyncDeviceService.java
@@ -52,7 +52,7 @@ public interface SyncDeviceService extends BaseService {
 	 * Never modify or reference this interface directly. Always use {@link SyncDeviceServiceUtil} to access the sync device remote service. Add custom service methods to {@link com.liferay.sync.service.impl.SyncDeviceServiceImpl} and rerun ServiceBuilder to automatically copy the method declarations to this interface.
 	 */
 	public SyncDevice registerSyncDevice(java.lang.String type,
-		int buildNumber, int featureSet, java.lang.String uuid)
+		long buildNumber, int featureSet, java.lang.String uuid)
 		throws PortalException;
 
 	/**
