+++ modules/apps/sync/sync-api/src/main/java/com/liferay/sync/service/SyncDeviceServiceUtil.java
@@ -42,7 +42,7 @@ public class SyncDeviceServiceUtil {
 	 * Never modify this class directly. Add custom service methods to {@link com.liferay.sync.service.impl.SyncDeviceServiceImpl} and rerun ServiceBuilder to regenerate this class.
 	 */
 	public static com.liferay.sync.model.SyncDevice registerSyncDevice(
-		java.lang.String type, int buildNumber, int featureSet,
+		java.lang.String type, long buildNumber, int featureSet,
 		java.lang.String uuid)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService()
